requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
unstructured[pdf]